TASK6

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ long int LONGINT;
 long long int LG;
 float FL;
 
-printf("Size of a char: %llu byte\n"), (unsigned long)sizeof(CHARTYPE);
-printf("Size of an int: %llu byte(s)\n"), (unsigned long)sizeof(INTTYPE);
-printf("Size of a long int: %llu byte(s)\n"), (unsigned long)sizeof(LONGINT);
-printf("Size of a long long int: %llu byte(s)\n"), (unsigned long)sizeof(LG);
-printf("Size of a float: %llu byte(s)\n"), (unsigned long)sizeof(FL);
+printf("Size of a char: %zu byte\n",sizeof(CHARTYPE));
+printf("Size of an int: %zu byte(s)\n",sizeof(INTTYPE));
+printf("Size of a long int: %zu byte(s)\n",sizeof(LONGINT));
+printf("Size of a long long int: %zu byte(s)\n",sizeof(LG));
+printf("Size of a float: %zu byte(s)\n",sizeof(FL));
 return (0);
 }
